+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/bisq_easy/offerbook/Filters.java
@@ -19,7 +19,7 @@ package bisq.desktop.main.content.bisq_easy.offerbook;
 
 import bisq.chat.ChatChannel;
 import bisq.chat.ChatMessage;
-import bisq.desktop.main.content.chat.message_container.list_view.ChatMessageListItem;
+import bisq.desktop.main.content.chat.message_container.list.ChatMessageListItem;
 import lombok.Getter;
 
 import java.util.function.Predicate;
