+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/bisq_easy/offerbook/MyOfferMessageBox.java
@@ -22,10 +22,10 @@ import bisq.chat.ChatMessage;
 import bisq.desktop.components.containers.Spacer;
 import bisq.desktop.components.controls.DropdownMenu;
 import bisq.desktop.components.controls.DropdownMenuItem;
-import bisq.desktop.main.content.chat.message_container.list_view.ChatMessageListItem;
-import bisq.desktop.main.content.chat.message_container.list_view.ChatMessagesListController;
-import bisq.desktop.main.content.chat.message_container.list_view.ChatMessagesListModel;
-import bisq.desktop.main.content.chat.message_container.list_view.message_box.BubbleMessageBox;
+import bisq.desktop.main.content.chat.message_container.list.ChatMessageListItem;
+import bisq.desktop.main.content.chat.message_container.list.ChatMessagesListController;
+import bisq.desktop.main.content.chat.message_container.list.ChatMessagesListModel;
+import bisq.desktop.main.content.chat.message_container.list.message_box.BubbleMessageBox;
 import bisq.i18n.Res;
 import de.jensd.fx.fontawesome.AwesomeIcon;
 import javafx.geometry.Insets;
