+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/bisq_easy/open_trades/MyProtocolLogMessageBox.java
@@ -21,8 +21,8 @@ import bisq.chat.ChatChannel;
 import bisq.chat.ChatMessage;
 import bisq.desktop.common.Icons;
 import bisq.desktop.components.controls.BisqTooltip;
-import bisq.desktop.main.content.chat.message_container.list_view.ChatMessageListItem;
-import bisq.desktop.main.content.chat.message_container.list_view.ChatMessagesListController;
+import bisq.desktop.main.content.chat.message_container.list.ChatMessageListItem;
+import bisq.desktop.main.content.chat.message_container.list.ChatMessagesListController;
 import bisq.network.p2p.services.confidential.ack.MessageDeliveryStatus;
 import de.jensd.fx.fontawesome.AwesomeDude;
 import javafx.geometry.Pos;
